LayoutTests/css3/calc/zoom-with-em-expected.txt

Tests that zooming a calc expression containing 'em' units works correctly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 308 is 308
PASS 309 is 309
PASS 310 is 310
PASS 310 is 310
PASS 310 is 310
PASS 310 is 310
PASS 310 is 310
PASS 310 is 310
PASS 310 is 310
PASS 310 is 310
PASS successfullyParsed is true

TEST COMPLETE
